+++ lib/model/wellness/WellnessRing.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class WellnessRingDefinition {