+++ lib/ui/canvas/CanvasCourseAnnouncementsPanel.dart
@@ -22,12 +22,14 @@ import 'package:flutter_html/flutter_html.dart';
 import 'package:illinois/model/Analytics.dart';
 import 'package:illinois/model/Canvas.dart';
 import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/Canvas.dart';
 import 'package:illinois/ui/canvas/CanvasAnnouncementDetailPanel.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
@@ -41,7 +43,7 @@ class CanvasCourseAnnouncementsPanel extends StatefulWidget with AnalyticsInfo {
   _CanvasCourseAnnouncementsPanelState createState() => _CanvasCourseAnnouncementsPanelState();
 }
 
-class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncementsPanel> {
+class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncementsPanel> with NotificationsListener {
   Map<int, List<CanvasDiscussionTopic>?>? _courseAnnouncementsMap;
   List<CanvasCourse>? _courses;
   int? _selectedCourseId;
@@ -51,9 +53,27 @@ class _CanvasCourseAnnouncementsPanelState extends State<CanvasCourseAnnouncemen
   @override
   void initState() {
     super.initState();
+    NotificationService().subscribe(this, [
+      AppDateTime.notifyTimeZoneChanged,
+    ]);
     _selectedCourseId = widget.courseId;
     _courses = Canvas().courses;
     _loadAnnouncements();
+  }
+
+  @override
+  void dispose() {
+    NotificationService().unsubscribe(this);
+    super.dispose();
+  }
+
+  // NotificationsListener
+
+  @override
+  void onNotification(String name, dynamic param) {
+    if (name == AppDateTime.notifyTimeZoneChanged) {
+      setStateIfMounted(() {});
+    }
   }
 
   @override