+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Time zone setting and date display based on the selected time zone [#5859](https://github.com/rokwire/illinois-app/issues/5859).
+- Added Campus Dining subsection to Dining [#5872](https://github.com/rokwire/illinois-app/issues/5872).
 ### Changed
 - Match Notifications buttons to match app style [#5868](https://github.com/rokwire/illinois-app/issues/5868).